refactor: split monolithic lib.rs into focused modules
- Split ~1000+ lines from lib.rs into 5 specialized modules:
  - src/error.rs: Error handling utilities
  - src/normalize.rs: Path normalization algorithms
  - src/prefix.rs: Existing prefix computation and symlink handling
  - src/symlink.rs: Symlink chain resolution
  - src/windows.rs: Windows-specific functionality

- Improved error reporting with path-aware error construction
- Enhanced symlink handling with better . and .. processing
- Restored fast-path optimization for non-existing first components
- Maintained full API compatibility and all 110+ tests pass

Version: 0.3.1
