+++ .github/copilot-instructions.md
@@ -1,0 +1,9 @@
+Please refer to `AGENTS.md` at the repository root for all operational guidance, coding and docs conventions, CI usage, API rules, and contribution practices.
+
+Key sections in `AGENTS.md`:
+- Local CI Parity (how to run CI locally on Windows/Unix/WSL)
+- Before Committing (commit message and staged-diff expectations)
+- Code Style (clippy, rustfmt, Rust API guidelines)
+- Documentation Guidelines (README/lib.rs/API docs alignment)
+- Path Handling Rules and API & Conversion Rules (security-critical)
+- Do / Don’t (for agents and automation)
