+++ .gitignore
@@ -15,3 +15,6 @@ Cargo.lock
 
 # Object files
 *.o
+
+# IDE
+.vscode
