+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "soft-canonicalize"
-version = "0.3.0"
+version = "0.3.1"
 edition = "2021"
 authors = ["David Krasnitsky <dikaveman@gmail.com>"]
 description = "Path canonicalization that works with non-existing paths."
@@ -8,7 +8,7 @@ license = "MIT OR Apache-2.0"
 repository = "https://github.com/DK26/soft-canonicalize-rs"
 homepage = "https://github.com/DK26/soft-canonicalize-rs"
 documentation = "https://docs.rs/soft-canonicalize"
-keywords = ["path", "canonicalize", "filesystem", "security", "symlink"]
+keywords = ["path", "canonicalize", "filesystem", "realpath", "symlink"]
 categories = ["filesystem", "os", "development-tools"]
 readme = "README.md"
 rust-version = "1.70.0"
