+++ src/symlink.rs
@@ -1,0 +1,88 @@
+use std::path::{Path, PathBuf};
+use std::{fs, io};
+
+use crate::error::error_with_path;
+use crate::normalize::simple_normalize_path;
+
+/// Maximum number of symlinks to follow before giving up.
+/// This matches the behavior of std::fs::canonicalize and OS limits:
+/// - Linux: ELOOP limit is typically 40
+/// - Windows: Similar limit around 63
+/// - Other Unix systems: Usually 32-40
+pub const MAX_SYMLINK_DEPTH: usize = if cfg!(target_os = "windows") { 63 } else { 40 };
+
+/// Resolve a symlink chain using read_link only (no extra metadata calls).
+/// Notes:
+/// - Uses a visited set on textual paths to detect cycles without extra IO
+/// - Caps depth at MAX_SYMLINK_DEPTH (or a smaller heuristic for common system symlinks)
+/// - Re-resolves relative symlink targets against the parent of the current link
+pub(crate) fn resolve_simple_symlink_chain(symlink_path: &Path) -> io::Result<PathBuf> {
+    let mut current = symlink_path.to_path_buf();
+    let mut depth = 0usize;
+    // Small Vec cycle detection avoids hashing overhead; chains are typically short.
+    let mut visited: Vec<std::ffi::OsString> = Vec::with_capacity(8);
+
+    // Heuristic: system symlinks are unlikely to be malicious chains; keep their budget small
+    let effective_max_depth = if is_likely_system_symlink(&current) {
+        5
+    } else {
+        MAX_SYMLINK_DEPTH
+    };
+
+    loop {
+        // Detect cycles using the textual path (no extra IO)
+        if visited.iter().any(|s| s == current.as_os_str()) {
+            return Err(error_with_path(
+                io::ErrorKind::InvalidInput,
+                symlink_path,
+                "Too many levels of symbolic links",
+            ));
+        }
+        visited.push(current.as_os_str().to_os_string());
+
+        match fs::read_link(&current) {
+            Ok(target) => {
+                depth += 1;
+                if depth > effective_max_depth {
+                    return Err(error_with_path(
+                        io::ErrorKind::InvalidInput,
+                        symlink_path,
+                        "Too many levels of symbolic links",
+                    ));
+                }
+
+                if target.is_absolute() {
+                    current = target;
+                } else if let Some(parent) = current.parent() {
+                    current = simple_normalize_path(&parent.join(target));
+                } else {
+                    current = target;
+                }
+            }
+            Err(_) => break, // not a symlink or cannot read; stop here
+        }
+    }
+
+    Ok(current)
+}
+
+/// Checks if a symlink is likely a system symlink that shouldn't consume depth budget
+#[cfg(target_os = "macos")]
+fn is_likely_system_symlink(path: &Path) -> bool {
+    let s = path.to_string_lossy();
+    s.starts_with("/var") || s.starts_with("/tmp") || s.starts_with("/etc")
+}
+
+#[cfg(target_os = "linux")]
+fn is_likely_system_symlink(path: &Path) -> bool {
+    let s = path.to_string_lossy();
+    s.starts_with("/lib")
+        || s.starts_with("/usr/lib")
+        || s.starts_with("/bin")
+        || s.starts_with("/sbin")
+}
+
+#[cfg(not(any(target_os = "macos", target_os = "linux")))]
+fn is_likely_system_symlink(_path: &Path) -> bool {
+    false
+}
